+++ app-misc/pinephonepro-manjaro-tweaks/files/60-libcamera-nodes-by-default.lua
@@ -1,0 +1,10 @@
+table.insert(libcamera_monitor.rules, {
+  matches = {
+    {
+      { "node.name", "matches", "libcamera_input.*" },
+    },
+  },
+  apply_properties = {
+    ["priority.session"] = 1000,
+  },
+})